md_pic_tools/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
pillow==8.3.2
pyperclip3==0.4.1
pywin32==301; platform_system == 'Windows'
